statis kdtre...

+++ src/models/mokp/mokp.h
@@ -39,7 +39,7 @@ double mokpnode_axis_val(MOKPNode *node, int h);
 typedef struct MOKPTree{
     MOKP *mokp;
     int n_nodes;
-    int n_comparisons;
+    long long n_comparisons;
 
     MOKPNode *root;
     MOKPNode *tail;
@@ -49,7 +49,7 @@ typedef struct MOKPTree{
 MOKPTree *mokptree_new(MOKP *mokp);
 
 /* Solving */
-int mokp_dynprog(MOKP *mokp, int use_kdtree, int k, int *idxs, int *n_comps);
+int mokp_dynprog(MOKP *mokp, int use_kdtree, int k, int *idxs, long long *n_comps);
 
 #endif
